requirements-serverless.txt
tensorflow-cpu==2.1.0
numpy==1.18.1
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0